chore: allow non-gov emails when inviting team members
